created http server
This is a simple HTTP server that listens on port 5000 and sends a response to the client. The server reads the client's request, prints it to the console, and sends a response back to the client. HOW to Create a Simple HTTP Server in Java : 1. Create a ServerSocket object and bind it to a port. 2. Wait for a client to connect using the accept() method. 3. Read the client's request using the getInputStream() method of the Socket class. 4. Send a response to the client using the getOutputStream() method of the Socket class. 5. Close the connection.